refactored code and console logs removed
