+++ test/smart-contract-tests/Parameter_AccountAddress.test.ts
@@ -6,7 +6,6 @@ import {
     InitContractPayload,
     ParameterType,
     SMParameter,
-    SMStruct,
 } from '../../src/types';
 import * as ed from 'noble-ed25519';
 import { getAccountTransactionSignDigest } from '../../src/serialization';
@@ -22,7 +21,7 @@ const senderAccountAddress =
 const wrongPrivateKey =
     'ce432f6cca0d47caec1f45739331dc354b6d749fdb8ab7c2b7f6cb24db39ca0c';
 // test case for init contract
-test('init contract with the wrong private key', async () => {
+test('Parameter of AccountAddress with the wrong private key', async () => {
     const nextAccountNonce = await client.getNextAccountNonce(
         new AccountAddress(senderAccountAddress)
     );
@@ -35,42 +34,17 @@ test('init contract with the wrong private key', async () => {
         sender: new AccountAddress(senderAccountAddress),
     };
 
-    const contractName = 'UserMixed';
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    const inputParams: SMParameter<SMStruct> = {
-        type: ParameterType.Struct,
-        value: [
-            {
-                type: ParameterType.U8,
-                value: 51,
-            } as SMParameter<number>,
-            {
-                type: ParameterType.String,
-                value: 'Concordium',
-            } as SMParameter<string>,
-            {
-                type: ParameterType.String,
-                value: 'Zug',
-            } as SMParameter<string>,
-            {
-                type: ParameterType.String,
-                value: 'Denmark',
-            } as SMParameter<string>,
-            {
-                type: ParameterType.Array,
-                value: {
-                    type: ParameterType.U8,
-                    value: [20, 30, 45],
-                },
-            },
-        ] as SMStruct,
+    const contractName = 'UserAddress';
+    const inputParams: SMParameter<AccountAddress> = {
+        type: ParameterType.AccountAddress,
+        value: new AccountAddress(senderAccountAddress),
     };
     const baseEnergy = 300000n;
 
     const initModule: InitContractPayload = {
         amount: new GtuAmount(0n),
         moduleRef: new ModuleReference(
-            'a68ff8d897ee06efb2782839c7bc530134bb6b32e9a96c081e027d20ddc5004d'
+            '3116f9792bf412e273d10a52bf42b78dddb9833d0c51c6cf79831f3b96866569'
         ),
         contractName: contractName,
         parameter: inputParams,
@@ -97,5 +71,6 @@ test('init contract with the wrong private key', async () => {
         initContractTransaction,
         signatures
     );
+
     expect(result).toBeTruthy();
 }, 300000);
