+++ test/smart-contract-tests/Parameter_Array_U8.test.ts
@@ -6,13 +6,10 @@ import {
     InitContractPayload,
     ParameterType,
     SMParameter,
-    SMStruct,
+    SMArray,
 } from '../../src/types';
 import * as ed from 'noble-ed25519';
-import {
-    getAccountTransactionHash,
-    getAccountTransactionSignDigest,
-} from '../../src/serialization';
+import { getAccountTransactionSignDigest } from '../../src/serialization';
 import { getNodeClient } from '../testHelpers';
 import { AccountAddress } from '../../src/types/accountAddress';
 import { GtuAmount } from '../../src/types/gtuAmount';
@@ -21,11 +18,11 @@ import { Buffer } from 'buffer/';
 import { ModuleReference } from '../../src/types/moduleReference';
 const client = getNodeClient();
 const senderAccountAddress =
-    '3gLPtBSqSi7i7TEzDPpcpgD8zHiSbWEmn23QZH29A7hj4sMoL5';
+    '4ZJBYQbVp3zVZyjCXfZAAYBVkJMyVj8UKUNj9ox5YqTCBdBq2M';
 const wrongPrivateKey =
-    '681de9a98d274b56eace2f86eb134bfc414f5c366022f281335be0b2d45a8988';
+    'ce432f6cca0d47caec1f45739331dc354b6d749fdb8ab7c2b7f6cb24db39ca0c';
 // test case for init contract
-test('init contract with the wrong private key', async () => {
+test('Parameter of Array of U8 with the wrong private key', async () => {
     const nextAccountNonce = await client.getNextAccountNonce(
         new AccountAddress(senderAccountAddress)
     );
@@ -38,31 +35,20 @@ test('init contract with the wrong private key', async () => {
         sender: new AccountAddress(senderAccountAddress),
     };
 
-    const contractName = 'MarkSheet';
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    const inputParams: SMParameter<SMStruct> = {
-        type: ParameterType.Struct,
-        value: [
-            {
-                type: ParameterType.U8,
-                value: 25,
-            } as SMParameter<number>,
-            {
-                type: ParameterType.U8,
-                value: 23,
-            } as SMParameter<number>,
-            {
-                type: ParameterType.U8,
-                value: 22,
-            } as SMParameter<number>,
-        ] as SMStruct,
+    const contractName = 'INDBankU83';
+    const inputParams: SMParameter<SMArray<number>> = {
+        type: ParameterType.Array,
+        value: {
+            type: ParameterType.U8,
+            value: [26, 27, 51],
+        },
     };
     const baseEnergy = 300000n;
 
     const initModule: InitContractPayload = {
         amount: new GtuAmount(0n),
         moduleRef: new ModuleReference(
-            '05b53729aef402a001226cd3d91e1044a4a6b29cd063b290f7798eb487750a8b'
+            'd883ce0619ff08711dea54f48acda749db62ace27d8dcc5eddc3796d5ebb5688'
         ),
         contractName: contractName,
         parameter: inputParams,
@@ -87,14 +73,8 @@ test('init contract with the wrong private key', async () => {
 
     const result = await client.sendAccountTransaction(
         initContractTransaction,
-
         signatures
     );
 
-    const txHash = await getAccountTransactionHash(
-        initContractTransaction,
-        signatures
-    );
-    console.log(txHash);
     expect(result).toBeTruthy();
 }, 300000);
