+++ test/smart-contract-tests/Parameter_I64.test.ts
@@ -22,7 +22,7 @@ const senderAccountAddress =
 const wrongPrivateKey =
     'ce432f6cca0d47caec1f45739331dc354b6d749fdb8ab7c2b7f6cb24db39ca0c';
 //test case for update contract
-test('update contract with the wrong private key', async () => {
+test('Parameter of I64 with the wrong private key', async () => {
     const nextAccountNonce = await client.getNextAccountNonce(
         new AccountAddress(senderAccountAddress)
     );
@@ -36,7 +36,6 @@ test('update contract with the wrong private key', async () => {
     };
 
     const receiveName = 'INDBankU83.insertAmount7';
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
     const inputParams: SMParameter<bigint> = {
         type: ParameterType.I64,
         value: BigInt(-2000000),
