+++ test/smart-contract-tests/Parameter_Struct_U8_3Variables.test.ts
@@ -3,10 +3,10 @@ import {
     AccountTransactionHeader,
     AccountTransactionSignature,
     AccountTransactionType,
-    UpdateContractPayload,
-    ContractAddress,
+    InitContractPayload,
     ParameterType,
     SMParameter,
+    SMStruct,
 } from '../../src/types';
 import * as ed from 'noble-ed25519';
 import { getAccountTransactionSignDigest } from '../../src/serialization';
@@ -15,14 +15,14 @@ import { AccountAddress } from '../../src/types/accountAddress';
 import { GtuAmount } from '../../src/types/gtuAmount';
 import { TransactionExpiry } from '../../src/types/transactionExpiry';
 import { Buffer } from 'buffer/';
-
+import { ModuleReference } from '../../src/types/moduleReference';
 const client = getNodeClient();
 const senderAccountAddress =
     '4ZJBYQbVp3zVZyjCXfZAAYBVkJMyVj8UKUNj9ox5YqTCBdBq2M';
 const wrongPrivateKey =
     'ce432f6cca0d47caec1f45739331dc354b6d749fdb8ab7c2b7f6cb24db39ca0c';
-//test case for update contract
-test('update contract with the wrong private key', async () => {
+// test case for init contract
+test('Parameter of Struct (3 varaibles) with the wrong private key', async () => {
     const nextAccountNonce = await client.getNextAccountNonce(
         new AccountAddress(senderAccountAddress)
     );
@@ -35,35 +35,43 @@ test('update contract with the wrong private key', async () => {
         sender: new AccountAddress(senderAccountAddress),
     };
 
-    const receiveName = 'INDBankU83.insertAmount1';
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    const inputParams: SMParameter<number> = {
-        type: ParameterType.U8,
-        value: 20,
+    const contractName = 'MarkSheet';
+    const inputParams: SMParameter<SMStruct> = {
+        type: ParameterType.Struct,
+        value: [
+            {
+                type: ParameterType.U8,
+                value: 25,
+            } as SMParameter<number>,
+            {
+                type: ParameterType.U8,
+                value: 23,
+            } as SMParameter<number>,
+            {
+                type: ParameterType.U8,
+                value: 22,
+            } as SMParameter<number>,
+        ] as SMStruct,
     };
-    const contractAddress = {
-        index: BigInt(108),
-        subindex: BigInt(0),
-    } as ContractAddress;
-    const baseEnergy = 30000n;
+    const baseEnergy = 300000n;
 
-    const updateModule: UpdateContractPayload = {
-        amount: new GtuAmount(1000n),
-        contractAddress: contractAddress,
-        receiveName: receiveName,
+    const initModule: InitContractPayload = {
+        amount: new GtuAmount(0n),
+        moduleRef: new ModuleReference(
+            '05b53729aef402a001226cd3d91e1044a4a6b29cd063b290f7798eb487750a8b'
+        ),
+        contractName: contractName,
         parameter: inputParams,
         maxContractExecutionEnergy: baseEnergy,
-    } as UpdateContractPayload;
+    } as InitContractPayload;
 
-    const updateContractTransaction: AccountTransaction = {
+    const initContractTransaction: AccountTransaction = {
         header: header,
-        payload: updateModule,
-        type: AccountTransactionType.UpdateSmartContractInstance,
+        payload: initModule,
+        type: AccountTransactionType.InitializeSmartContractInstance,
     };
 
-    const hashToSign = getAccountTransactionSignDigest(
-        updateContractTransaction
-    );
+    const hashToSign = getAccountTransactionSignDigest(initContractTransaction);
     const signature = Buffer.from(
         await ed.sign(hashToSign, wrongPrivateKey)
     ).toString('hex');
@@ -74,7 +82,7 @@ test('update contract with the wrong private key', async () => {
     };
 
     const result = await client.sendAccountTransaction(
-        updateContractTransaction,
+        initContractTransaction,
         signatures
     );
