+++ test/smart-contract-tests/Parameter_Struct_U8_String_String.test.ts
@@ -6,12 +6,10 @@ import {
     InitContractPayload,
     ParameterType,
     SMParameter,
+    SMStruct,
 } from '../../src/types';
 import * as ed from 'noble-ed25519';
-import {
-    getAccountTransactionHash,
-    getAccountTransactionSignDigest,
-} from '../../src/serialization';
+import { getAccountTransactionSignDigest } from '../../src/serialization';
 import { getNodeClient } from '../testHelpers';
 import { AccountAddress } from '../../src/types/accountAddress';
 import { GtuAmount } from '../../src/types/gtuAmount';
@@ -20,11 +18,11 @@ import { Buffer } from 'buffer/';
 import { ModuleReference } from '../../src/types/moduleReference';
 const client = getNodeClient();
 const senderAccountAddress =
-    '3gLPtBSqSi7i7TEzDPpcpgD8zHiSbWEmn23QZH29A7hj4sMoL5';
+    '4ZJBYQbVp3zVZyjCXfZAAYBVkJMyVj8UKUNj9ox5YqTCBdBq2M';
 const wrongPrivateKey =
-    '681de9a98d274b56eace2f86eb134bfc414f5c366022f281335be0b2d45a8988';
+    'ce432f6cca0d47caec1f45739331dc354b6d749fdb8ab7c2b7f6cb24db39ca0c';
 // test case for init contract
-test('init contract with the wrong private key', async () => {
+test('Parameter of Struct (U8, string, string variables) with the wrong private key', async () => {
     const nextAccountNonce = await client.getNextAccountNonce(
         new AccountAddress(senderAccountAddress)
     );
@@ -37,18 +35,31 @@ test('init contract with the wrong private key', async () => {
         sender: new AccountAddress(senderAccountAddress),
     };
 
-    const contractName = 'UserAddress';
+    const contractName = 'UserDetails';
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    const inputParams: SMParameter<AccountAddress> = {
-        type: ParameterType.AccountAddress,
-        value: new AccountAddress(senderAccountAddress),
+    const inputParams: SMParameter<SMStruct> = {
+        type: ParameterType.Struct,
+        value: [
+            {
+                type: ParameterType.U8,
+                value: 51,
+            } as SMParameter<number>,
+            {
+                type: ParameterType.String,
+                value: 'Concordium',
+            } as SMParameter<string>,
+            {
+                type: ParameterType.String,
+                value: 'Zug',
+            } as SMParameter<string>,
+        ] as SMStruct,
     };
     const baseEnergy = 300000n;
 
     const initModule: InitContractPayload = {
         amount: new GtuAmount(0n),
         moduleRef: new ModuleReference(
-            '3116f9792bf412e273d10a52bf42b78dddb9833d0c51c6cf79831f3b96866569'
+            '8e4222c5e7d7a2f950aac1f8073b35d43a8ee11108e52534470a70c57aa6f780'
         ),
         contractName: contractName,
         parameter: inputParams,
@@ -73,14 +84,7 @@ test('init contract with the wrong private key', async () => {
 
     const result = await client.sendAccountTransaction(
         initContractTransaction,
-
         signatures
     );
-
-    const txHash = await getAccountTransactionHash(
-        initContractTransaction,
-        signatures
-    );
-    console.log(txHash);
     expect(result).toBeTruthy();
 }, 300000);
