+++ test/smart-contract-tests/Parameter_U8.test.ts
@@ -22,7 +22,7 @@ const senderAccountAddress =
 const wrongPrivateKey =
     'ce432f6cca0d47caec1f45739331dc354b6d749fdb8ab7c2b7f6cb24db39ca0c';
 //test case for update contract
-test('update contract with the wrong private key', async () => {
+test('Parameter of U8 with the wrong private key', async () => {
     const nextAccountNonce = await client.getNextAccountNonce(
         new AccountAddress(senderAccountAddress)
     );
