--- test/smart-contract-tests/initContractWithParametersStringArray.test.ts
@@ -1,81 +1,0 @@
-import {
-    AccountTransaction,
-    AccountTransactionHeader,
-    AccountTransactionSignature,
-    AccountTransactionType,
-    SMArray,
-    InitContractPayload,
-    ParameterType,
-    SMParameter,
-} from '../../src/types';
-import * as ed from 'noble-ed25519';
-import { getAccountTransactionSignDigest } from '../../src/serialization';
-import { getNodeClient } from '../testHelpers';
-import { AccountAddress } from '../../src/types/accountAddress';
-import { GtuAmount } from '../../src/types/gtuAmount';
-import { TransactionExpiry } from '../../src/types/transactionExpiry';
-import { Buffer } from 'buffer/';
-import { ModuleReference } from '../../src/types/moduleReference';
-const client = getNodeClient();
-const senderAccountAddress =
-    '4ZJBYQbVp3zVZyjCXfZAAYBVkJMyVj8UKUNj9ox5YqTCBdBq2M';
-const wrongPrivateKey =
-    'ce432f6cca0d47caec1f45739331dc354b6d749fdb8ab7c2b7f6cb24db39ca0c';
-// test case for init contract
-test('init contract with the wrong private key', async () => {
-    const nextAccountNonce = await client.getNextAccountNonce(
-        new AccountAddress(senderAccountAddress)
-    );
-    if (!nextAccountNonce) {
-        throw new Error('Nonce not found!');
-    }
-    const header: AccountTransactionHeader = {
-        expiry: new TransactionExpiry(new Date(Date.now() + 3600000)),
-        nonce: nextAccountNonce.nonce,
-        sender: new AccountAddress(senderAccountAddress),
-    };
-
-    const contractName = 'INDBankStringArray';
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    const inputParams: SMParameter<SMArray<string>> = {
-        type: ParameterType.Array,
-        value: {
-            type: ParameterType.String,
-            value: ['Concordium', 'Zug', 'CCD'],
-        },
-    };
-    const baseEnergy = 300000n;
-
-    const initModule: InitContractPayload = {
-        amount: new GtuAmount(0n),
-        moduleRef: new ModuleReference(
-            'edff6bce3cc9117d88db66d55f43695a0ee337e6803b48fb1bfa7ef76ec98447'
-        ),
-        contractName: contractName,
-        parameter: inputParams,
-        maxContractExecutionEnergy: baseEnergy,
-    };
-
-    const initContractTransaction: AccountTransaction = {
-        header: header,
-        payload: initModule,
-        type: AccountTransactionType.InitializeSmartContractInstance,
-    };
-
-    const hashToSign = getAccountTransactionSignDigest(initContractTransaction);
-    const signature = Buffer.from(
-        await ed.sign(hashToSign, wrongPrivateKey)
-    ).toString('hex');
-    const signatures: AccountTransactionSignature = {
-        0: {
-            0: signature,
-        },
-    };
-
-    const result = await client.sendAccountTransaction(
-        initContractTransaction,
-        signatures
-    );
-
-    expect(result).toBeTruthy();
-}, 300000);
